Time: 31 ms (59.28%), Space: 95.8 MB (16.02%) - LeetHub

+++ 0128-longest-consecutive-sequence/0128-longest-consecutive-sequence.java
@@ -1,15 +1,15 @@
 class Solution {
     public int longestConsecutive(int[] nums) {
-        HashSet<Long> hs = new HashSet<>();
-        for(long x : nums){
+        HashSet<Integer> hs = new HashSet<>();
+        for(int x : nums){
             hs.add(x);
         }
         int max = 0;
 
-        for(long x: hs){
+        for(int x: hs){ // important i forgotten to write hs here i wrote nums
             if(!hs.contains(x-1)){
                 int length = 1;
-                long cnt = x;
+                int cnt = x;
 
                 while(hs.contains(cnt+1)){
                     length++;
